requirements.txt
gradio~=4.44.1
jwt~=1.3.1
PyJWT~=2.10.1
bcrypt~=4.2.1
pymongo~=4.10.1
python-dotenv~=1.0.1
requests~=2.32.3